remove csrf.exempt from api.action, update the JS calls to include CSRF header
