+++ ckan/public/base/javascript/modules/confirm-action.js
@@ -100,6 +100,13 @@ this.ckan.module('confirm-action', function (jQuery) {
         var form = this.el.closest('form');
       }
 
+      // get the csrf value
+      var csrf_value = $('meta[name=_csrf_token]').attr('content')
+      // set the hidden input
+      var hidden_csrf_input = $('<input name="_csrf_token" type="hidden" value="'+csrf_value+'">')
+      // insert the hidden input at the beginning of the form
+      hidden_csrf_input.prependTo(form)
+
       form.appendTo('body').submit();
     },
