+++ ckan/public/base/javascript/modules/follow.js
@@ -30,19 +30,6 @@ this.ckan.module('follow', function($) {
 			this.el.on('click', this._onClick);
 		},
 
-		/* Adds the X-CSRFToken header
-		*/
-		addCsrfTokenHeader: function() {
-			var csrftoken = $('meta[name=_csrf_token]').attr('content')
-			$.ajaxSetup({
-				beforeSend: function(xhr, settings) {
-					if (!/^(GET|HEAD|OPTIONS|TRACE)$/i.test(settings.type) && !this.crossDomain) {
-						xhr.setRequestHeader("X-CSRFToken", csrftoken)
-					}
-				}
-			})
-		},
-
 		/* Handles the clicking of the follow button
 		 *
 		 * event - An event object.
@@ -51,7 +38,6 @@ this.ckan.module('follow', function($) {
 		 */
 		_onClick: function(event) {
 			var options = this.options;
-			this.addCsrfTokenHeader()
 			if (
 				options.action
 				&& options.type
