+++ ckan/views/api.py
@@ -15,7 +15,6 @@ from werkzeug.datastructures import MultiDict
 
 import ckan.model as model
 from ckan.common import json, _, g, request, current_user
-from ckan.config.middleware.flask_app import csrf
 from ckan.lib.helpers import url_for
 from ckan.lib.base import render
 from ckan.lib.i18n import get_locales_from_config
@@ -208,8 +207,6 @@ def _get_request_data(try_url_params: bool = False):
 
 # View functions
 
-# exempt the CKAN API actions from csrf-validation.
-@csrf.exempt
 def action(logic_function: str, ver: int = API_DEFAULT_VERSION) -> Response:
     u'''Main endpoint for the action API (v3)
 
@@ -238,7 +235,7 @@ def action(logic_function: str, ver: int = API_DEFAULT_VERSION) -> Response:
         u'auth_user_obj': current_user
     })
     model.Session()._context = context
-
+ 
     return_dict: dict[str, Any] = {
         u'help': url_for(u'api.action',
                          logic_function=u'help_show',
